+++ apps/demo-forms-granularity/CHANGELOG.md
@@ -1,5 +1,14 @@
 # demo-forms-granularity
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @nexus-state/core@0.2.0
+  - @nexus-state/react@0.2.0
+  - @nexus-state/devtools@0.2.0
+
 ## 0.0.1
 
 ### Patch Changes