+++ apps/demo-middleware/CHANGELOG.md
@@ -1,5 +1,14 @@
 # demo-middleware
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @nexus-state/core@0.2.0
+  - @nexus-state/react@0.2.0
+  - @nexus-state/middleware@1.0.3
+
 ## 0.0.1
 
 ### Patch Changes